Program to find file using bash scripting
write a Bash script that searches the user's Documents/ folder for a file name "ese2025.txt"; if the file does not exist, create the file with string "Found First!". If the file exists, append the string "Found Again!" to the file. End your script by sending the contents of the ~/Document/ese2025.txt file to standard output.